feat: сделано изменение групп пользователя
