+++ client/gql/dcis/fragments/period_fields.graphql
@@ -1,3 +1,5 @@
+#import "./period_group_fields.graphql"
+
 fragment PeriodFields on PeriodType {
   id
   name
@@ -12,6 +14,9 @@ fragment PeriodFields on PeriodType {
     src
     size
     __typename
+  }
+  periodGroups {
+    ...PeriodGroupFields
   }
   __typename
 }
