+++ client/gql/dcis/mutations/period/add_period_group.graphql
@@ -1,0 +1,15 @@
+#import "../../fragments/period_group_fields.graphql"
+
+mutation AddPeriodGroup($name: String!, $periodId: ID!) {
+  addPeriodGroup(input: {name: $name, period: $periodId}) {
+    success
+    errors {
+      messages
+      field
+      __typename
+    }
+    periodGroup {
+      ...PeriodGroupFields
+    }
+  }
+}
