+++ client/gql/dcis/mutations/period/change_period_group_privileges.graphql
