+++ client/gql/dcis/mutations/period/change_user_period_groups.graphql
@@ -1,0 +1,22 @@
+#import "../../fragments/period_group_fields.graphql"
+#import "../../../core/fragments/user_fields.graphql"
+
+mutation ChangeUserPeriodGroups($userId: ID!, $periodGroupIds: [ID!]!) {
+  changeUserPeriodGroups(input: {
+    userId: $userId,
+    periodGroupIds: $periodGroupIds
+  }) {
+    success
+    errors {
+      field
+      messages
+      __typename
+    }
+    periodGroups {
+       ...PeriodGroupFields
+    }
+    user {
+      ...UserFields
+    }
+  }
+}
