+++ client/gql/dcis/mutations/period/copy_period_groups.graphql
@@ -1,4 +1,4 @@
-#import "../../../core/fragments/user_fields.graphql"
+#import "../../fragments/period_group_fields.graphql"
 
 mutation CopyPeriodGroups($periodId: ID!, $selectedPeriodId: ID!, $periodGroupIds: [ID!]!) {
   copyPeriodGroups(input: {
@@ -13,21 +13,7 @@ mutation CopyPeriodGroups($periodId: ID!, $selectedPeriodId: ID!, $periodGroupId
       __typename
     }
     periodGroups {
-      id
-      name
-      createdAt
-      users {
-        ...UserFields
-        __typename
-      }
-      privileges {
-        id
-        name
-        key
-        createdAt
-        __typename
-      }
-      __typename
+      ...PeriodGroupFields
     }
   }
 }
