+++ client/gql/dcis/mutations/period/delete_period_group.graphql
