--- client/gql/dcis/mutations/project/add_period_group.graphql
@@ -1,29 +1,0 @@
-#import "../../../core/fragments/user_fields.graphql"
-
-mutation AddPeriodGroup($name: String!, $periodId: ID!) {
-  addPeriodGroup(input: {name: $name, period: $periodId}) {
-    success
-    errors {
-      messages
-      field
-      __typename
-    }
-    periodGroup {
-      id
-      name
-      createdAt
-      users {
-        ...UserFields
-        __typename
-      }
-      privileges {
-        id
-        name
-        key
-        createdAt
-        __typename
-      }
-      __typename
-    }
-  }
-}
