--- client/gql/dcis/mutations/project/change_period_group_users.graphql
@@ -1,19 +1,0 @@
-#import "../../../core/fragments/user_fields.graphql"
-
-mutation ChangePeriodGroupUsers($periodGroupId: Int!, $usersIds: [ID!]) {
-  changePeriodGroupUsers( input: {
-    periodGroupId: $periodGroupId,
-    usersIds: $usersIds
-  }) {
-    success
-    errors {
-      field
-      messages
-      __typename
-    }
-    users {
-      ...UserFields
-      __typename
-    }
-  }
-}
